Remove 'hoist.ts`
